+++ CHANGELOG.md
@@ -6,4 +6,7 @@ minor code | changes
 ---------- | -------
 0.10       | obstest outputs examples.  Extended noise models (using Webb 1998 and Wolin&McNamara.  Add obstest.yaml documentation
 0.10.post1 | change manifest so that examples are uploaded to PyPI
-
+0.10.post2 | Fixed bugs in selecting unique seed_ids and
+           | processing max_days > data_days, added npz and spectrogram outputs
+0.10.post3 | Added basic unittests
+0.11       | Fixed bug in plotPSDs if no match between inventory and data seed_ids